requirements.txt
tensorflow==2.0.0
pygame
numpy
pillow
